moved smartmobsystem to errorgate folder
